Moved files into proper folders for easy way to find correct files
