+++ ActiviGoApi/ActiviGoApi.Services/DTOs/BookingDTOs/BookingReadDTO.cs
@@ -1,4 +1,4 @@
-namespace ActiviGoApi.Services.DTOs
+namespace ActiviGoApi.Services.DTOs.BookingDTOs
 {
     public class BookingReadDTO
     {
