+++ ActiviGoApi/ActiviGoApi.Services/DTOs/BookingDTOs/BookingUpdateDTO.cs
@@ -1,6 +1,6 @@
 using System.ComponentModel.DataAnnotations;
 
-namespace ActiviGoApi.Services.DTOs
+namespace ActiviGoApi.Services.DTOs.BookingDTOs
 {
     public class BookingUpdateDTO
     {
